Only process the first touch in drag pan gestures
There was a bug where we would still arm the drag pan gesture's timer,
even when the touch was not the first one. This would sometimes result
in the recognizer being in a "dirty" state and recognizing the gesture
innapropriately.

Before: Pinching would frequently put the recognizer in the dirty state,
now pinching no longer affects the state of the recognizer